finatra-http: Remove reference to AsyncStream in StreamingResponse
Problem

StreamingResponse holds a hard reference to the AsyncStream you are streaming.
This means that the whole stream is kept in memory until the streaming response
is complete. Infinite streams result in Out of Memory Exceptions.

Solution

StreamingResponse now takes only by-name parameters for the AsyncStream, and
stores a function which will generate the stream when called. This avoids
keeping a reference and allows the head of the stream to be garabage collected
during writing.

Result

Infinite streaming is possible without running out of memory.

RB_ID=890205