finatra-http: Request in request case classes no longer requires Inject annotation
Problem

The guice and finatra @Inject annotations are overloaded. They are required in
order to get Request in a request case class. However, the Request is not
provided via the Injector. This causes confusion when you try to @Inject the
Request elsewhere.

Solution

No longer require the @Inject in request case classes and remove the previously
deprecated @RequestInject to prevent confusion.

Result

Request can be included in a request case class and should not use an
annotation. This makes it clear it is not being injected using guice.

RB_ID=888197
TBR=true