finatra/http: Deprecate Netty 3 Cookie Method in EnrichedResponse
Problem

One version of the cookie method in ResponseBuilder's EnrichedResponse
class takes a Netty 3 Cookie as a parameter. Netty types should not be
exposed as part of Finatra's public API.

Solution

Deprecate the particular instance of the cookie method and encourage
folks to use the one which takes a Finagle HTTP cookie instead.

RB_ID=888683