jackson: Handle joda exception in FinatraCaseClassDeserializer
Problem

parseConstructorValues method doesn't not handle joda
IllegalFieldValueException, this causes that the exception is
re-thrown leading to 'Internal Server Error' while deserializing
a request into a case class.

This happens when the case class have a LocalDate field and the
request argument is an invalid LocalDate.

Solution

Handle joda IllegalFieldValueException to not return an
'Internal Server Error' with meaningless feedback but 'Bad request'
having enough feedback for the client.