finatra-http: Decompose the ThrowableExceptionMapper for easier customization
Problem

It's not currently possible to customize the general
`c.t.finatra.http.internal.exceptions.ThrowableExceptionMapper` because its
visibility is hidden (as we did not want to publicize an implementation we
were not convinced would not change). Thus, if users want to develop their
own exception mapper over a Throwable type they have to start from scratch
which typically means repeating a lot of the `ThrowableExceptionMapper`.

Solution

Decompose the `ThrowableExceptionMapper` such that if you want to customize
it's function you can register a mapper specifically over the type for
customization and not have to replicate the rest of the
`ThrowableExceptionMapper` as it's functionality is now split into specific,
distinct, individually registered exception mappers.

Result

Users can more specifically replace just the functionality for Throwable
exceptions handled by the default
`c.t.finatra.http.internal.exceptions.ThrowableExceptionMapper` to customize
it's exception-to-response handling for their purposes.

RB_ID=891666
TBR=true