finatra-http: Allow controllers to forward requests to another route
Problem

There is no way for a controller to forward a request to another route
(possibly defined in another controller). Only support for a request
redirect exists today, which entails the client issuing a new request
to the desired route. Forwarding will re-use the original request.
Several other frameworks have this concept and it would be useful to
support this in Finatra.

Solution

Add a way for Controllers to forward the handling of a request to
another route. Forwarded requests will not be re-played through any
server defined filter chain. However, any Controller-defined filter
on the "forwarded-to" route will be executed on the forwarded request.

Result

Users are able to define routes which can forward to other routes
re-using the original request.

RB_ID=883224