finatra-http: Support for Controller route prefixes
Problem

It would be useful to allow for scoping of a set of routes underneath
a defined prefix.

Solution

Allow a way for users to define a prefixed block wherein all routes
defined will be prepended with the prefix value. E.g.,

```
class MyController extends Contoller {

  get("/foo") { request: Request =>
    "bar"
  }

  prefix("/1.1") {
    get("/foo") { request: Request =>
      "bar"
    }
  }
}
```

Would yield two routes:

```
GET    /foo
GET    /1.1/foo
```

How the prefix is provided to the Controller is up to the user and
gives the flexibility of allowing it to be injected or hardcoded or
computed in some manner.

Result

A declarative way to define routes underneath a given prefix.

RB_ID=894695