Bump the specs2 version to something cross-published for 2.11 and 2.12.
Summary:
Problem

We were using an older copy of specs2 which wasn't cross-published for
scala 2.12. Additionally, it had been necessary to shade that copy.

Solution

Upgrade to a version of specs2 that is cross-published for 2.11 and 2.12,
and which (helpfully) no longer requires the shading.

Test Plan: https://ci.twitter.biz/job/source_sandbox_phab_production/424/

Reviewers: mbilotti, ity, qma, mlandis, ccoco, O890 source:/finatra/

Reviewed By: ccoco, O890 source:/finatra/

Subscribers: #rb_devprod-build-team, #rb_csl

Differential Revision: https://phabricator.twitter.biz/D17624