inject-app: Fix TestInjector to properly parse flags
Problem

The TestInjector is incorrectly parsing flags when no value
has been provided. This incorrectly set the default of
Flaggable[T] instead of the flag's set default which is a
problem for Boolean types.

Solution

In the TestInjector#parseFlags method we do not execute any
logic in the branch where the moduleFlag has not provided
value. The flags will be correctly created when installing
the modules via InstalledModules#create.

Result

Correct handling of defaulted Boolean flags in modules when
using the TestInjector.

RB_ID=901525