finatra-http: Support for optional trailing slashes in HTTP routes
Problem

We don't currently elegantly support a way for users to define routes
that are agnostic to a trailing slash. E.g., define a route, "/groups"
that would match a request URI of both "/groups" and "/groups/".

Solution

Allow for users to specify routes that end with "/?" which denotes the
route allows for an optional trailing slash. Update the code accordingly.

Result

Users can define routes that support optional trailing slashes if they
so desire.

RB_ID=893167
TBR=true