finatra: Update CHANGELOG for JUnitRunner removal and specs2 update
Problem

Update the finatra/CHANGELOG.md for updates landed in a previous commit:

We were using an older copy of specs2 which wasn't cross-published for
scala 2.12. Additionally, it had been necessary to shade that copy.

Solution

Upgrade to a version of specs2 that is cross-published for 2.11 and 2.12,
and which (helpfully) no longer requires the shading. Additionally, depend
on core and mock modules independently, rather than a merged artifact.
Finally, remove the JUnit RunWith annotation from finatra abstract test
classes, since it is no longer necessary with pants, and was never necessary
with sbt.

RB_ID=902337