finatra-http: Add Java support for ExceptionMappers
Problem

We don't currently expose the ExceptionMapper functionality for easy use
from Java in terms of creating and registering a new custom mapper in
Java.

Solution

Create an `AbstractExceptionMapper` and update the
`c.t.finatra.http.routing.HttpRouter` to allow for registration of
`AbstractExceptionMappers`.

Result

ExceptionMappers can be created and registered in Java with Java
HttpServers.

RB_ID=902995