finatra-http: Update EmbeddedTwitterServer admin route matching logic
Problem

The last update to allow routes which start with "/admin" to be exposed
on a server's external interface and routes which do not start with
"/admin" to be exposed on the admin interface introduced an interesting
corner-case where a user could choose to add two routes with different
HTTP verbs, one on the external interface and one on the admin interface.

The EmbeddedHttpServer does not currently distinguish by HTTP method when
deciding if a request should be shunted to the embedded admin http client.
It currently only uses path and thus would route a request to the external
endpoint incorrectly to the admin interface.

This work pointed out an omission in the translation from Finatra Routes
to TwitterServer AdminHttpServer.Routes where the HTTP method was not
carried over to the admin route.

Solution

Update the EmbeddedHttpServer#chooseHttpClient to take into account the
HTTP method of the incoming request when trying to determine if the request
should be routed to the embedded admin http client. Fix TwitterServer
AdminHttpServer.Route creation to properly include the HTTP Method.

Result

More correct embedded server request routing logic. Correct route
semantics in the TwitterServer AdminHttpServer created routes.

RB_ID=905887
TBR=true