finatra-http: Add Java support for declaring admin routes
Problem

We don't currently support adding admin routes nor route indexes
in the AbstractController meant for usage from Java.

Solution

Add overloaded methods which wil allow for Java users to specify
if a route is an admin route and for configuring a RouteIndex.

Result

Ability for Java users to add admin routes and potentially include
them in the admin index.

RB_ID=906264
TBR=true