finatra: Move userguide to sphinx
Problem

Right now the userguide is tied to creation and deployment of the
entire Finatra site. However, we want to be able to create and publish
blog posts independently of updating the userguide. Much like the
current scaladoc publishing.

Solution

Move the userguide to be able to be generated by sbt and thus broken
out into a separate process for deployment. Additionally update the
pushsite script to allow for pushing the entire site (i.e., when doing
a release) or just the parts outside of the userguide and scaladocs,
(i.e., for publising a new blog post or page).

Result

Easier process to create documentation.

RB_ID=906380
TBR=true