finagle-thrift, finatra-thrift: Fix support for inherited BaseServiceIface implementations
Problem

When BaseServiceIface implementations that extend other services are
converted into a MethodIface and served we don't currently instantiate
the correct class as the case is not found. Instead we end up
instantiating the superclass and serving that service as the endpoint
which can lead to errors of the form:

`org.apache.thrift.TApplicationException: Invalid method name: 'ping'`

when attempting to make an RPC call to a function defined in the child
service as only the parent is being served. The primary user of this
use case is the Finatra framework thrift support.

Solution

Extend the `tryScroogeFinagleService` method in `rich.scala` to also
take into account a class in the form of `Service$MethodIface` which
is the class type for BaseServiceIface services which extends other
services.

Result

Finatra thrift servers can serve thrift services which extend another
service.

RB_ID=907608