Add support for single-level nesting through companion objects
Problem

The fix for "case classes nested one level inside of objects" didn't work if
that object was a companion object, since earlier code would "find" the
companion object's class instead of the object itself.

Solution

If we're trying to find an case-class-in-object, don't attempt to use the
object's companion class.

Result

This now works for a single level of nesting through companion objects.

Signed-off-by: Christopher Coco <ccoco@twitter.com>

RB_ID=908606