finatra-http: Increase composability and flexibility of RouteDSL
Problem:

Today, the routing DSL supports two composable pieces (prefix and
filter). These are implemented under different paradigms; filter
produces a new Route context while prefix takes a block and uses a var
to track the prefix.

There are a few issues / unexpected behaviors that result:
* Prefixes are not composable
* Filtered routes inside of prefixes don't work (https://github.com/twitter/finatra/issues/392)
* Filtered prefixes are not supported
* The syntax is rigid between the two

Solution

The changes here revist the structure of RouteDSL and unify the way
routes are defined in Controllers to increase flexibility and remove
unexpected behavior and bugs. Namely, mutable state is now centralized
to a twitter-util Var cell and an API is laid out for creating
"specialized" RouteDSL contexts (FilteredDSL and PrefixedDSL).

This allows for a great amount of flexibility for the user when defining
routes as well as increasing the extensibility of the DSL for future
additions. For example:

Filtered prefixes:

filter[MyFilter].prefix("/v1/prefix") {
    // ...
}

Cascading prefixes:

prefix("/v1") {
  prefix("/foo") {
    prefix("/bar") {
     // ..
    }
  }
}

"Freestyle" filter syntax:

filter[MyFilter] {
  // ...
}