inject-core: Allow for binding of higher-kinded types in testing utilities
Problem

We'd like to replace the functionality of `@Bind` with a usage that does not
suffer from the caveats listed in the current documentation:
https://twitter.github.io/finatra/user-guide/testing/index.html#id1

Especially, the requirement that test members which access the embedded server
be defined as lazy or within a test method since this has repeatedly proven to
be subtle and error-prone for users.

We previously introduced a `bind[T]` DSL in the testing utilities but it is limited
by its use of Scala Manifests and thus higher-kinded type information is lost. This
API has advantages over using `@Bind` in that:
1. it is less error-prone as it does not have the lazy declaration requirement
   for fields that access the embedded server or TestInjector.
2. it is specified per object graph (i.e., per-server or per-TestInjector) instead
   of per test file and thus allows for binding different instances of like types
   for multiserver tests.
3. the API forces the user to specify the bound type thus alleviating the issue
   of incorrect type inference.

Solution

Update the `#bind[T]` testing DSL to work with Scala TypeTags and provide parity
across the `EmbeddedTwitterServer`, `EmbeddedHttpServer`, `EmbeddedThriftServer`,
`EmbeddedApp`, and the `TestInjector` such that users can bind instances of
a given type and instances of a general type with a binding annotation including
higher-kinded types.

Deprecate the `@Bind` mechanisms.

Note, there is a breaking API change in regards to the TestInjector. Adding this
support in the TestInjector now requires that users call the `TestInjector#create`
method in order to build the injector and that this is done *after* any calls to
`TestInjector#bind`. Previously, an `Injector` was directly returned from
`TestInjector#apply` which is no longer true.

Result

A better API for replacing bound instances in an object graph when testing that
is more straight-forward and less error prone.

RB_ID=910011
TBR=true