finatra-http: Correctly support adding Java AbstractController by instance
Problem

Java support does not correctly handle adding an AbstractController instance
to the HttpRouter (rather than adding it by type).

Solution

Correctly support the case(s) where AbstractControllers can be added by passing
an instance to the HttpRouter to call the `AbstractController#configureRoutes`
method to configure the routes in the AbstractController.

RB_ID=910502
TBR=true