finatra-http: Make JsonPatchExceptionMapper public
Problem

The json patch documentation[1] describes how to set up twitter-server
to make use of json patch requests. The docs explain that one must register the
JsonPatchExceptionMapper, but outside of the `http` package this is not possible
as the class has a `private[http]` modifier.

[1]: https://twitter.github.io/finatra/user-guide/http/requests.html#json-patch-requests

Solution

Remove the `private[http]` modifier of JsonPatchExceptionMapper making it
possible to import and use.

Signed-off-by: Bryce Anderson <banderson@twitter.com>

RB_ID=910951