finatra-jackson: Support parsing Long value passed as String in FinatraDatetimeDeserializer
Problem

When using a custom case class Request with a date time injectable field
it is currently not possible to pass the value in millis.

Solution

Patch the FinatraDatetimeDeserializer to attempt to convert a given string
to a long if possible before constructing a DateTime object.

Result

Users can specify and use a custom case class Request like so:

```
case class ClassWithQueryParamDateTime(
  @QueryParam dateTime: DateTime)

...

get("/millis") { r: ClassWithQueryParamDateTime =>
  r.dateTime.getMillis
}
```

And a request will be parsed accordingly:

```
===========================================================================
HTTP GET /millis?date_time=1489719177279
[Header]	Host -> 127.0.0.1:60580
===========================================================================
2017-03-17 03:11:29,951 INF            AccessLoggingFilter       127.0.0.1 - - [17/Mar/2017:03:11:29 +0000] "GET /millis?date_time=1489719177279 HTTP/1.1" 200 13 601 "-"
---------------------------------------------------------------------------
[Status]	Status(200)
[Header]	Content-Type -> text/plain; charset=utf-8
[Header]	Server -> Finatra
[Header]	Date -> Fri, 17 Mar 2017 03:11:29 GMT
[Header]	Content-Length -> 13
1489719177279
```

RB_ID=911162