inject-server: Allow the process to end when any Awaitable is ready
Problem

ListeningServer exposes an Awaitable interface, but we sometimes want to kill
our servers not only when the server dies, but also on other conditions.
Await.all requires that we're able to track down and kill every one of the
things we're awaiting, which is not always the right thing to do.

Solution

Wait for any of the awaitables we wait on in finatra to die to let the process
die.

RB_ID=911028