finatra-http: Ensure underlying members are injected for AbstractControllers
Problem

When adding an AbstractController to the HttpRouter, underlying
members are not injected. This means the ResponseBuilder is not
set in the controller.

Solution

Make sure when an AbstractController is added to the HttpRouter
that any underlying members are injected.

Result

Java users will regain access to the ResponseBuilder for creating
responses from Java controllers.

RB_ID=911635
TBR=true