inject-app: Run installed modules postInjectorStartup before server function
Problem

There is a disconnect in the lifecycle order of the postInjectorStartup calls.
The installed modules postInjectorStartup function is currently called after
the server, meaning the server's router is configured before the module's
singleton startup is performed. Additionally, framework modules are currently
just concatenated to the end of the list of modules, when their functionality is
increasing expected to be overridden, swapped out, or otherwise augmented by user
defined behavior. Neither of these are ideal and currently make it hard to reason
about the lifecycle of certain events.

Solution

Ensure framework modules are added before any other modules and run the installed
modules postInjectorStartup before calling the server's function.

Result

Reasoning about the server lifecycle will be a bit more straight-forward and
things like the exception manager logic for adding and overridding mappers becomes
simpler and functions more as intended.

RB_ID=911965
TBR=true