finatra-http: Return a JsonParseException when JSON is not parsable as custom case class request object
Problem

When a user defines an HTTP controller route using a custom case class request object
as the callback input type, if the route receives JSON that is not parseable into the
case class we should return a JsonParseException which will be surfaced as a BadRequest
by the default exception mappers.

Solution

Update the FinatraCaseClasssDeserializer to properly handle this case.

RB_ID=912529