source: Migrate Finagle off HTTP codec (use HTTP stack instead)
Disclaimer: Half of this (if not more) is done by Moses Nakamura.

CSL is in the middle of [migrating Finagle HTTP from Netty 3 to Netty 4][1] and we
want to make sure **all users** will be migrated no matter what API they use to
configure HTTP clients/servers. Because of that we move services using HTTP codec
(became package private as part of this RB) to HTTP stack such that they will join
the Netty 4 migration train.

[1]: https://groups.google.com/a/twitter.com/forum/#!topic/finaglers/aijSf9ZRiCw

RB_ID=912427
TBR=true
NO_USER_HOOK=1