finatra: Change the constructor parameter of HttpClientException
Problem

When using the HttpClient.executeJson method, if the status of the response
differs from expectedStatus, an HttpClientException is returned. The status and
contentString of the response are used when creating an HttpClientException, but
the value of status is actually inaccessible.

Solution

Change the constructor parameter of HttpClientException.

Signed-off-by: Yufan Gong <yufang@twitter.com>

RB_ID=913964
TBR=true