finatra/http: Add support for json patch array indexing
Problem

According to the JSON patch specification RFC 6902, arrays should be patchable
using the JSON patch operations.

RFC 6901 (referenced by RFC 6902) describes the special json pointer /- which
points to the last element when retrieving a node, or a new element at the end
of the array when creating a node.

Solution

Extend the JsonPatchOperator class with further logic for array indexing.
Matching on node type (ArrayNode or ObjectNode) before deciding how to access
the next child node, handling invalid and out of bounds indices.

Signed-off-by: Yufan Gong <yufang@twitter.com>

RB_ID=913279