finatra-jackson: Fix Either Type Deserialization in Scala 2.12
Problem

In Scala 2.12, the `Either` type extends the `Product` trait. This breaks
the case class deserializer type check in
`FinatraCaseClassDeserializers`.

Solution

Add an explicit condition for the `Either` type, so that the case class
deserializer is not selected and the `EitherDeserializer` is selected
instead.

Result

An `Either` object can be converted to JSON and back using the
`FinatraObjectMapper` in both Scala 2.11 and Scala 2.12.

Signed-off-by: Ryan O'Neill <ryano@twitter.com>

RB_ID=917699