finatra-http: Add close hook for StreamingResponse
Summary: Problem

An async stream may be backed by resources that must be released but
the only mechanism for doing this is consuming the entire stream which
cannot be guaranteed and may not be efficient. This can result in
resource leaks when using the finatra StreamingResponse.

Solution

Allow passing a Closable alongside the AsyncStream when constructing
a StreamingResponse to give an explicit handle for cleaning up resources.

JIRA Issues: CSL-4738

Differential Revision: https://phabricator.twitter.biz/D64013