[strato] Create (but not yet use) a new live-idl BUILD target
Summary: To start using the new scrooge powered liveidls, we'll need a BUILD target
that depends on all the *-strato thrift targets that we care about. This is
similar to the one currently powering our live idl command (though that one
depends on the *-scala variants of these targets).

Because we chose to be strict about strato namespaces in .thrift files, I had to:
1) Make create_thrift_libraries not generate for strato by default
2) update all the create_thrift_libraries we care about (transitively) to include strato bindings
3) Add a strato namespace to all relevant .thrift files

For future reference, all thrift files we depend on can be found via:
./pants filedeps strato/config:liveidl-bin | grep "\.thrift"

This only creates the BUILD target + fixes all the upstream BUILD and .thrift files.

I wanted to keep this (very long, mechanical) diff separate from the one that actually
changes our unit tests to use this target, so I will send that in a follow up review.
I'd also like to get this in before I start to hit merge conflicts as it touches a lot of files.

NOTE: `./pants bundle strato/config:liveidl-bin` now takes about 6 seconds after it's been run once (after build cache has kicked in), and that's generating about 2k .strato files instead of the 120 that we currently do. (we may want to go back to filtering out the structs we don't need, 2k is a lot of .strato files).

NOTE: probably the only file worth looking at in this diff is: strato/config/BUILD
link: https://phabricator.twitter.biz/D63899#ee6d6b43

TBR=true

Differential Revision: https://phabricator.twitter.biz/D63899