finatra/jackson: Add @SnakeCaseMapper binding annotation
Summary: Problem

It's currently possible to default a FinatraObjectMapper to snake case, while
injecting a camel case mapper where required by using the @CamelCaseMapper
annotation. Unfortunately the reverse is not possible.

Solution

Add a new Guice provides method to FinatraJacksonModule that allows for the
explicit injection of a FinatraObjectMapper that will use the SNAKE_CASE
property naming strategy.

Result

Use the new @SnakeCaseMapper annotation to request injection of a
FinatraObjectMapper configured with the SNAKE_CASE property naming strategy.

Differential Revision: https://phabricator.twitter.biz/D77698