finatra-http: Remove all stream head references with jsonarray generation
Summary: Problem

The StreamingResponse incorrectly holds on to references to the head
of the stream which can result in a reference leak as objects are
never garbage collected.

Solution

Remove all stream head references in the generation of the jsonarray.

Result

Less chance of a reference leak and more performant code.

Signed-off-by: Christopher Coco <ccoco@twitter.com>

JIRA Issues: CSL-4999

Differential Revision: https://phabricator.twitter.biz/D84891