inject-server: Add ability to fail embedded server startup on lint rule violation
Summary: Problem

There exists a class of lint rules which are evaluated at server startup.
It is not currently possible to fail a test if a lint rule violation is
detected upon server startup.

Solution

Add a flag in the embedded servers that when set to true will fail
server startup if a lint rule violation is detected. This will then fail
the test running.

Result

Allow users to write tests which fail if any linting rules are in
violation after server startup.

JIRA Issues: CSL-5081

Differential Revision: https://phabricator.twitter.biz/D82399