finatra-http: Remove the dependency on bijection-util
Summary: Problem

finatra-http has a dependency on bijection-util to convert Scala Futures to
Twitter Futures. However bijection-util depends on an older version of
util-core which may become incompatible with the version Finatra uses.

Solution

Introduce a simple implicit conversion from a Scala Future to a Twitter
Future for use in the CallbackConverter. Remove the dependency on
bijection-util.

Result

Less chance of an issue stemming from a diamond dependency around util-core.

JIRA Issues: CSL-5148

Differential Revision: https://phabricator.twitter.biz/D86640