inject-server: Only run linting when failOnValidation is true
Summary: Problem

There is currently not a good story around isolating a server
in the canonical usage of FeatureTests in which the server
is specified as a top-level val in the test. This means that
it is hard for users to get actionable lint warnings.

Solution

Only run the linting on the embedded server when the
failOnLintValidation flag is set to true until we have a
better way to isolate the server startup from shared state
pollution.

JIRA Issues: CSL-5152

Differential Revision: https://phabricator.twitter.biz/D87065