finatra-jackson: Fix issue swallowing IllegalArgumentExceptions from Validations
Summary: Problem

A catch clause in the
`c.t.finatra.json.internal.caseclass.jackson.FinatraCaseClassDeserializer` is
too broad as it catches thrown `IllegalArgumentException`s from field
validations when the annotation is applied to a field of the incorrect type,
e.g., when `@Max` is applied to a String field.

Solution

The genesis of the catch clause is around an exception that occurs when
a java LocalDate cannot be properly constructed from an incoming JSON
value. The thought was to be more general in what was caught here but
in this case we actually need to keep the specificity to avoid swallowing
exceptions thrown by Validations attempted on unsupported types for the
validator.

JIRA Issues: CSL-5284

Differential Revision: https://phabricator.twitter.biz/D95306