inject-server: Ensure EmbeddedTwitterServer has started to close httpAdminClient
Summary: Problem/Solution

When the embedded server closes before fully starting
(e.g., due to an error) we attempt to close the created
admin http client which fails if the server has not started
obscuring the real error. Only attempt to close the
admin http client if the server has recorded that it has
successfully started.

JIRA Issues: CSL-5396

Differential Revision: https://phabricator.twitter.biz/D111294