finagle-thrift, finagle-thriftmux: Remove `ReqRep` specific methods
Summary: Problem

Exposing two different APIs for `ServicePerEndpoint` and `ReqRepServicePerEndpoint`
in the ThriftRichClient, meant that any APIs built on-top also had to carry
this bifurcation, in some cases making it challenging to provide a clear
API. Since the only difference between the two are the Service types, it should be
possible to use a single API over differing types.

Solution

The main cause of the bifurcation was that were two non-related builders defined,
one the `ServicePerEndpointBuilder` and one the `ReqRepServicePerEndpointBuilder`.

These builers are now related with the `ReqRepServicePerEndpointBuilder` being
a subclass (a more specialized version) of the `ServicePerEndpointBuilder`.

This allows us to express an API purely in terms of the parent-type and in turn
makes having specific "ReqRep" methods unncessary, thus we remove them.

Differential Revision: https://phabricator.twitter.biz/D123341