inject-thrift-client: `ServicePerEndpointModule` -> `ThriftMethodBuilderClientModule` and update `ThriftClientModule`
Summary: Problem

The introduction of the `c.t.inject.thrift.modules.ServicePerEndpointModule`
(now the `c.t.inject.thrift.modules.ThriftMethodBuilderClientModule`),
gives users a way to configure a ThriftMux client with per-method semantics.

While it is simple to not perform any configuration per-method to generate
a simple client the construction of the module requires type information
that is not strictly necessary. Additionally, the `ServicePerEndpointModule`
only provides bindings to a `MyService.ServicePerEndpoint` and a
`MyService.MethodPerEndpoint` which wraps the `MyService.ServicePerEndpoint`.

If a user wants to create a simple ThriftMux client to a Thrift service
with no per-method configuration or filtering, the `ThriftMethodBuilderClientModule`
can be considered overkill.

Solution

Un-deprecate the `c.t.inject.thrift.modules.ThriftClientModule` and update
it for parity in terms of creating and configuring the underlying ThriftMux
client and swith to ThriftMux support *only*. Add documentation on it's
intended use. THe `ThriftClientModule` is typed to a single interface param
and exposes the configured client as a bound instance of that interface type.

Supported types are dictated by the `c.t.finagle.thrift.ThriftRichClient`.

We also update the `c.t.inject.thrift.modules.ServicePerEndpointModule` to
the more consistently named: `c.t.inject.thrift.modules.ThriftMethodBuilderClientModule`

Result

User's have a choice of APIs when constructing a ThriftMux client in Finatra:
they can use the more flexible per-method `ThriftMethodBuilderClientModule`
or the simpler `ThriftClientModule` to just express ThriftMux client configuration
with no per-method semantics and no filtering of the exposed client interface.

TBR=true

Differential Revision: https://phabricator.twitter.biz/D129891