inject-thrift-client: Register to close created Thrift client interfaces
Summary: Problem

Finatra provides utilities to create Thrift client interfaces
but does not register them to be closed upon graceful exit
of an application.

Solution

Update the `c.t.inject.modules.ThriftClientModule` and the
`c.t.inject.modules.ThriftMethodBuilderClientModule` to register a function
which will await on closing of the provided Thrift client interface.

Wire in this callback to be performed in an `onExit` block of
the `c.t.inject.app.App`.

JIRA Issues: CSL-5941

Differential Revision: https://phabricator.twitter.biz/D138266