inject-server: Add TwitterServer#setup lifecycle method for PUBSUB
Summary: Problem/Solution

For services which need to create and await on sometype of Publisher or
Subscriber is it not clear where in the lifecycle this should happen. There
are currently two places we recommend, `postInjectorStartup` and the more
common, `postWarmup`. These both come with the caveat that users need
to remember to call the `super` version in both cases and their names are
very non-obvious for the use case.

Introduce `TwitterServer#setup` which is called in the `postInjectorStartup`
lifecycle method with the specific intention of being the correct place
in the server lifecycle where users can "setup" publishers and subscribers.

JIRA Issues: CSL-5966

Differential Revision: https://phabricator.twitter.biz/D135827