inject-server: Add a lint rule in `c.t.inject.server.TwitterServer#warmup`
Summary: Problem

Services can experience a negative impact to their success rate if they do not
implement a warmup function for their server. That is, if specific code paths
are not exercised before accepting traffic to the server.

It is generally considered a good practice for servers to implement at least
some basic server warmup functionality.

Solution

To bring visbility to not having defined server warmup functionality, we implement
a lint rule in the base implementation of `c.t.inject.server.TwitterServer#warmup`
which always fails if the method is not overridden with a different implementation.

Because overrides which needlessly call `super.warmup` would trigger the lint rule
violation, we also clean up any stray calls to `super.warmup` in correctly overridden
implementations.

Result

If a server does not override the default implementation of `TwitterServer#warmup`
a lint rule violation will appear on the lint page of the HTTP admin interface.

JIRA Issues: CSL-5925

TBR=true

Differential Revision: https://phabricator.twitter.biz/D141267