finatra-http: Log ReaderDiscarded exceptions in StreamingResponse as info
Summary: Problem

When `StreamingResponse` is sending a response and the client disconnects, the
writer fails to write to the client, resulting in an exception and a warning
with a stacktrace sent to the logs. This is an expected exception and while
warning may be appropriate, a stacktrace is overkill and pollutes the logs.

Solution

Log `info` without a stacktrace when `ReaderDiscarded` is thrown, and log
`debug` with a stacktrace when other exceptions are thrown.

Result

Logs when clients disconnect from `StreamingResponse` servers will be less
noisy.

Differential Revision: https://phabricator.twitter.biz/D141453