inject-server: Capture errors on close of the underlying TwitterServer
Summary: Problem

We want to be able to capture failures on close of the underlying TwitterServer
in the embedded testing utilities.

Solution

a. Add a `EmbeddedTwitterServer#closeOnExit` which simplifies collecting Closables
to exit on shutdown of the server.
b. Update the `EmbeddedTwitterServer#close` to properly capture an exception from
closing the underlying TwitterServer. Capture the exception to make it available for
assertion.
c. Add `EmbeddedTwitterServer#assertCleanShutdown` which inspects for any
Throwable captured from closing the underlying server which it will then throw.

Result

Users are able to better assert the behavior on close of the underlying server.

JIRA Issues: CSL-6004, CSL-6008, CSL-3794

Differential Revision: https://phabricator.twitter.biz/D148946