finatra-http: Correctly handle Failures in FailureExceptionMapper
Summary: Problem

The `c.t.finatra.http.internal.exceptions.FailureExceptionMapper`
was split out from the then "FinatraDefaultExceptionMapper` some time
ago and wired in as a framework default exception mapper over the `Failure`
type. When this was split out it was updated to properly handle cases of
Failures which are not "wrapped".

Solution

Update the `FailureExceptionMapper` to properly handle (and not rethrow) non
wrapped Failures. Update the logging of "unhandled" failures to attempt to
map to the defined level in the Failure to reduce noisiness of logs.

JIRA Issues: CSL-6253

Differential Revision: https://phabricator.twitter.biz/D158591