finatra-http: Fix FinagleMDCAdapter to initialize properly
Summary: Problem

We lazily initialize the mutable map backing the FinagleMDCAdapter which
is stored in a `Local`. When that initialization happens in a scope which is
popped when returning up the stack, the user loses any values stored in the
map. E.g., it is a problem if the user does something like this:

```
getUserById(id).map { user: User =>
  MDC.put("user_id", user.id)
  logger.info("Looked up user.)
  user
}.flatMap { user =>
  getUserPreferences(user)
}.flatMap { preferences =>
  logger.info("Checking preferences to send notification") // the "user_id" is not present here
  // as the backing map was only created in map closure above and has fallen out of scope
  if (preferences.sendNotification) {
    sendNotification(preferences)
  } else Future.Done
}
```

In the above code, by the time the second `logger` statement is executed
the MDC values are lost as they were lazily created in the `MDC.put` in
the `#map` and have been popped from the Local context.

Generally, this has worked as users typically add one of the MDC logging
filters, e.g., `TraceIdMDCFilter` which incidentally creates the backing
MDC map early enough in the stack. But this only works incidentally.

Solution

We need to initialize the mutable map eagerly to ensure that the backing
structure is eagerly available in the `Local`.

Result

The backing MDC map is created in the `LoggingMDCFilter` which does
the MDC initialization and thus is as early in the MDC stack as possible
when means the backing map implementation should be correctly available
when users transition across Future boundaries.

JIRA Issues: CSL-6267

TBR=true

Differential Revision: https://phabricator.twitter.biz/D159536