finatra-http: Improve Http Nack error handling
Summary: Problem

The `HttpNackFilter` converted NACK exceptions into generic
`HttpExceptions` which could be handled by a registered
`HttpExceptionMapper`. However, it is fairly common for users
to register their own `HttpExceptionMapper` and it is very likely
that users would have no idea that they need to account specially
for NACK exceptions.

Solution

We update to emit a specific `HttpNackException` from the
`HttpNackFilter` and provide a default exception mapper:
`HttpNackExceptionMapper` over that exception type which is
distinct from the `HttpExceptionMapper`.

Users can still override this mapper if they so choose but it
is specific to NACKs and therefor should be clearer as to the
behavior that is being changed.

Result

If users provide a new `ExceptionMapper[HttpException]` they will
not have to also account for dealing with NACK exceptions. Instead
if users want to customize or change the default HTTP NACK
handling they can register a `ExceptionMapper[HttpNackException]`
to replace the default supplied by the framework. In this way, users
do not lose NACK handling by just replacing the default `HttpException`
exception mapper.

JIRA Issues: CSL-6366

Differential Revision: https://phabricator.twitter.biz/D172456