inject-logback: Update AsyncAppender to also register wrapped appenders
Summary: Problem

We'd like to also list the appenders the AsyncAppender is wrapping in
the registry and update to ensure that we close any gauges when stopping
the AsyncAppender.

Solution

Update to include wrapped Appenders as a comma-separated string in the
registry in an "appenders" entry under the AsyncAppender entry. Capture
and close gauges in the `AsyncAppender#stop` method. Add tests for
the registry.

JIRA Issues: CSL-6473

Differential Revision: https://phabricator.twitter.biz/D177042