finatra: Add filter and route/method information to Registry
Summary: Problem

We want to expose more information about the Filters
and Controllers that compose a service.

Solution

Add information about the Filters, HTTP routes, and
Thrift methods to the registry for introspection on
what Filters and what Controllers make up a service.

Result

The registry can be inspected for information on the
Filters and Controllers that comprise a service.

JIRA Issues: CSL-6429

Differential Revision: https://phabricator.twitter.biz/D177583