finatra: Parallelize TravisCI building
Summary: Problem

The current TravisCI build is taking too long as caching of directories
appears to be broken (or now only caching upon successful builds).

However, we rely on the caching to kick in to eventually allow the
build to successfully complete.

Solution

We've tried a few times to parallelize the build in the same manner
as the Finagle TravisCI but previous attempts have failed. I'm hopeful
this one will work and we'll have a successful build.

I've updated the .travis.yml config to break out each Finatra project
into a matrix.

Differential Revision: https://phabricator.twitter.biz/D181848