inject-utils: Add 'toLoggable' implicit from Array[Byte] to String
Problem

I have an app that is reading messages that are encoded in
Thrift, and has been unable to decode some of those messages.  I want a way to
save those bytes off so that I can debug them later.  I want to log information
about those bytes(their size, and their hash) so that I can tell if this
is a problem with big messages, small messages, or the same message over and
over(because the hash is the same).

Solution

Add a new implicit `toLoggable` method onto Array[Byte] that will
convert the byte array to a String with the length of the byte array, a sha1
hash of the byte array, and the byte array base64 encoded.

Result

println(Array(0x1, 0x2, 0x3, 0x4).map(_.toByte).toLoggable)
"Length: 4, Sha1: 12dada1fff4d4787ade3333147202c3b443e376f, Base64: AQIDBA=="

Differential Revision: https://phabricator.twitter.biz/D182262